tawread fix Tin No

+++ GAZT/GAZT/ViewModel/NewDesignViewModel/CustomServicesViewModels/TransactionReceptionViewModel.cs
@@ -76,8 +76,8 @@ namespace EGAZT.ViewModel.NewDesignViewModel.CustomServicesViewModels
         string registrationNo;
         public string RegistrationNo { get { return registrationNo; } set { registrationNo = value; RaisePropertyChanged(); } }
 
-  int tIN;
-        public int TIN { get { return tIN; } set { tIN = value; RaisePropertyChanged(); } }
+  long tIN;
+        public long TIN { get { return tIN; } set { tIN = value; RaisePropertyChanged(); } }
 
         bool isTinNoVisible;
         public bool IsTinNoVisible { get { return isTinNoVisible; } set { isTinNoVisible = value; RaisePropertyChanged(); } }
@@ -411,7 +411,7 @@ namespace EGAZT.ViewModel.NewDesignViewModel.CustomServicesViewModels
 
                         if (!String.IsNullOrEmpty(data.TINNumber))
                         {
-                            TIN =int.Parse(data.TINNumber);
+                            TIN =long.Parse(data.TINNumber);
                             IsTinNoVisible = true;
                         }
